split item into components
