+++ src/components/Item/components/Blessings.css
@@ -1,0 +1,7 @@
+.blessing {
+  display: flex;
+}
+
+.blessing-rarity {
+  margin-right: 1em;
+}
