+++ src/components/Item/components/Divider.tsx
@@ -1,0 +1,3 @@
+export function Divider() {
+  return <hr className="MuiDivider-root MuiDivider-fullWidth css-pj146d" />
+}
