+++ src/components/Item/components/Title.css
@@ -1,0 +1,3 @@
+.item-title {
+  margin-bottom: 0.5em;
+}
