+++ src/components/Item/components/Title.tsx
@@ -1,0 +1,15 @@
+import type { ReactNode } from "react"
+
+export function Title({
+  children,
+  style,
+}: {
+  children: ReactNode
+  style?: React.CSSProperties
+}) {
+  return (
+    <div className="item-title" style={style}>
+      {children}
+    </div>
+  )
+}
