+++ src/components/Item/consts.ts
@@ -1,0 +1,15 @@
+export const raritySymbol = {
+  1: "Ⅰ",
+  2: "Ⅱ",
+  3: "Ⅲ",
+  4: "Ⅳ",
+  5: "Ⅴ",
+} as const
+
+export const traitRarityToRating = {
+  1: 25,
+  2: 35,
+  3: 45,
+  4: 55,
+  5: 65,
+} as const
